NEXUS-3204, fixed HttpClientProxyUtil to allow for null logger to be passed in, (it shouldn't require any logger, but that is a different issue)
Added a stub of a UT that which requires auth (this could/should) have been an IT, it would have been easier to just use a canned config and kick off a get method. It would have been a little slower, but would have been 3-4 lines of code to do the same thing. 

Added notes in RemoteBrowserResource about changing parent to AbstractResourceStoreContentPlexusResource (to simplify it)

git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@5870 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab