Realized that DisableShutdownHookMangler does not work reliably, since Jetty is unable to reliably get rid of the ShutDownThread after it's configured.
What happens is that we first configure the Server instance using Nexus' jetty.xml, that contains explicit line <Set name="stopAtShutdown">true</Set>. The setter invocation triggers the thread registration with JVM Runtime. Later, we invoked DisableShutdownHookMangler that does the opposite: server.setStopAtShutdown( false ), and shutdown hook unregistration happens, but a leak is already created since thread instance stays dangling.

Now, we modify the jetty.xml _before_ we use it for configuration. This is the cleanest, since we completely _reuse_ the same jetty.xml as in Nexus bundle, only we modify this one setting, making it suitable for embedded use. In bundle mode, this setting _is_ required.