More model separation, all pushed out into reusable modules, but...
The ITs still expect to have ALL of Nexus Core on classpath, since it brings up complete NexusConfiguration component to perform configuration validation.... bad. So, this work -- even if it started to "trim down" the IT classpath and did succeed (and buys a lot for example for nexus-rest-client-java), did not help for ITs to trim their class path down. Bummer. But, as a bright side, now our models are really separated out!