Adding a logback.xml default logback configuration file to nexus-test-common module.
This module is used in all OSS and Pro modules, and having the logback configuration included
it here, it will become available in all UTs and ITs, and will use INFO level for console logging.
Currently, we relied on logback default behaviour, and it defaulted to DEBUG level, created huge
amounts of console outputs just for nothing.

If a developer needs DEBUG log, the ${test.loglevel} system property key needs to be set to
"DEBUG" value (warning: surefire and failsafe does fork, so mvn -Dtest.loglevel is not enough!)