NEXUS-4536: fixing the main cause.
With NEXUS-4300 implemented, UIDLocks were "put into one". Before, there was
a separate lock for content and another one for attributes. As we were in the middle
of moving attributes into LS, this worked, but in last moment we backed out from this.
But alas, the fact that there is one -- now shared -- lock for both content and
attributes got forgot about. This caused a contention, that is more visible with "bigger"
files, since during a request a lastRequest attribute is updated, yielding exclusive
lock against same resource, that is potentially already shared-locked by another
thread that is in the middle of download. This resulted in congestions, and also
in artifact serving serialization, that was not very visible with "quick" (ie. smaller)
downloads, but is very visible with few hundred MB sized files. Simply put,
while you download it, nobody else cannot get it, all the requests will simply remain
blocked, until download finishes, and one of the waiting ones will be the next one
able to download.

Related changes:
6d0475e8552fbfa2440a923df33b7b3b6bd7eed5
83975d547947650f698c27d2adf495404ccce94d