NEXUS-4520: Fix for "${nexus-work}" appear if nexus-mock folder after running tests.
This is not in direct relation to NEXUS-4520 actually, but was fixed.
What was happening, that Mock module boots nexus in a "in situ" way (as opposed to
ITs where "cleanroom" boot happens in isolated classloader). This is needed to be able
to "mock" plexus resources properly. But this also means, that classloader executing
the test is having a "priority" over actual nexus' webapp classloader (jetty term).
But, due to a ahem-load of dependencies, almost all of Nexus is "lifted" into Test
UT. Hence, plexus-ehcahce, used by Security is there also, and it uses classloader
of it's own classes to locate the ehcache config on classpath. This makes
WEB-INF/classes invisible to it (it's in child classloader, not in test classloader),
hence creates a tmp folder for ehcahce using wrong configuration.

Fix for now is simply "lifting" the proper ehcache.xml config file into test
resources, but ultimately, the plexus-ehcache needs an update.