[NEXUS-4530] Wait for repair tasks to finish before running tests
Intermittent failure of DefaultIndexerManagerTest#testRepoKeywordSearch was caused by the fact that before running the test, remote url of "apache-snapshots" repository was updated. This triggered an reindex task to repair the index of apache snapshots. The reindex task runs at undefined pint in time. In case that the reindex task was started and did not finished before the test was run, the reindexing started by the test will simply bypass the indexing of apache snapshots as it is considered that reindexing will anyhow be performed. when this happen the search results will not contain artifacts from apache snapshots, so the test will fail.

Signed-off-by: Alin Dreghiciu <adreghiciu@gmail.com>