Do not run long running tests as part of default pre push profile
Renamed all tests (its) that are running for longer then 15 seconds(ish) to LRTest (as in Long Running Test) and excluded them from default profile (prepush).
The tests are still run on another profiles as in "it" profile.

On my machine 2.3 i7/quad it reduced the default build from 16 minutes to about 7.

Signed-off-by: Alin Dreghiciu <adreghiciu@gmail.com>