o [NEXUS-4600] Drop FileLock for artifactory mapping file
Problem here is that there are two webapps consuming the mapping
configuration:
* nexus
* artifactory-bridge

They have different classloaders, so the ReentrantLock precaution is not
enough to rule out the JVM trying to get the configurations FileLock
twice, leading to sporadic OverlappingFileLockExceptions.

FileLock is dropped completely because nexus should be the only one
reading the migration config.