NEXUS-4521: more fixes for leak spotted during test debug.
If IO error occurs in the middle of body transfer, the connection
would be left dangling until HttpClient CM timeouts on it. On heavily
loaded systems this might lead to connection pool depletion. Now
the method.releaseConnection() will be invoked as quick as possible.

Also, in the UT the config change was not applied explicitly.

This are still not the solutions, see comments on NEXUS-4521 issue
about setSoTimeout() method.