Fixing the UT fail.
The UT started failing since the AbstractNexusTask is now decoupled from ErrorReportingManager, and uses eventing. OTOH,
the newly added "receiver" is an EventInspector, but in this UT it was not even activated, since EventInspectorHost was
not activated at all. So, all the change to this test (despite reformatting of the sloppy formatted source) is only in
last test method, where we "awake" Nexus component that installs the EventInspectorHost, and implemented a bit change
in doCall() method to really wait for all async tasks to execute (since this test actually goes to real remote JIRA to
create issues for real, and it takes long).