NEXUS-4676: Upgrade strategy.
A bit involved. The goal is it, that IF legacy folder is detected, AND it's not
marked as upgraded (by placing a special file in there), a thread is spawned
(not using scheduler to avoid config changes) that is daemon (to not interfere
with possible reboots related to upgrade) that forces upgrade of all repositories
threads. The "mark" is placed per-repository and "globally" once if all the work
is done. This is needed to cover all the cases of upgrades:

pre-2.0 -> 2.0
pre-2.0 -> post-2.0
post-2.0 -> post-2.0

And also a variations on presence of a repository in upgraded instance, but to be
resistant on added repository AFTER upgrade (ie. while upgrade is being performed).