NEXUS-4735: fixes and more
The initial problem seems fixed, but I went over improving ITs, and it turned out, that
many of ITs had side effects it was misusing. Also, one complete set of ITs was never working (!)
but they were simply "passing" as it was timing out after 10 minutes or run!
Also, plexus was bounced per-IT, while there was no need for it at all (all the "IT plexus" is used
is to look up some helper clients and so).