NEXUS-4744: fix
The original NPE was not due to load, but due to config event and how workingFolder was
handled. As it turns out, having AttributeStorage components listen for config events
(and doing whatever they did) was totally nonsense, since there are no means that
a config event (related to nexus.xml changes) would change the nexus-work folder!
Doing the latter is possible only by bouncing Nexus. Hence, scrubbed.

Moreover, the way how Nexus decides is "transitioning" (for attributes) needed or not is
much smarter now with meaningful log messages. Before, we would waste CPU cycles (if not
actual IO) just to invoke methods on "legacy" storage that would simply return doing nothing.
Now, we detect the need for "transitioning", and on new instances we even don't install
any of those components, thus, not wasting any cycles on them.