NEXUS-4735: fix
This is the real culprit of the issue originally reported. Logback logfiles are released when logback instance
is GCed. This is obviously not seen in "normal" ITs since they log in separate place, but in this case, immediately
after a shutdown, next IT run failed since -- probably -- the GC of previous one did not ran/finish/whatever,
but we wanted immediately to start another IT that begins with deleting the folder. Again, why we never see this
in case of other ITs is that every IT logs to it's own "specific" place, not affected by "clean workdir" step of
preparations for IT run.