NXCM-3651: fix for host
The fix for EventInspectorHost that caps and makes sure Nexus instance on heavy load
will not increase cached threads in host pool endlessly (leading to OOM probably), but
will keep it capped. Also, when pool is full, caller threads will start to execute the
async inspectors.

This fix does NOT solve the indexer, when the threads were actually detected ("seen")
to stockpile up.