NXCM-3642: a followup.
A proper finishing of the change in decd41b64c6515a8822b248dc970c3bcb204faaf

That change above "exposed" a method doing "nfc purge only", but it kept the
"dual nature" of the expireCaches() method in case of proxy repositories.
Now the two methods are clearly and properly separated.