o [NXCM-3688] Add config to TextFields to decode html entities in the value
In the REST request payload, characters are html encoded if needed. They are
not decoded in the response to rule out XSS problems.

This leads to weird behavior for fields where you edit actual plain
text, for example in the repository name. If you save '"name"' it's
escaped to '&quot;name&quot;' on the server and sent back like this.
On the next save, the value '&quot;name&quot;' will be sent to the
server, and is escaped again.

This commit unescapes html entities in the value *displayed in the
TextField*, so that the original value is sent to the server again.