NEXUS-4218: NXCM-3698: fix
The implementation of NEXUS-4218 was buggy, and was triggered by
userAgent-less invocation by smart proxy causing it to download (in case
of proxy repository) the file not being present in cache, ie. to pre-populate
the proxy cache.

This was never detected before SP, because invoking "expireCaches" method was
possible only over UI, and you could not invoke it against paths not present in
cache -- since UI would not represent them.