NEXUS-4770: Multiple fixes for Walker implementation.
Major issue reported in jira above was the presence of reportWalkEnd() method
in the catch of the inner block, that throws _wrapped_ exception that was catched
by the outer catch block. This was resulting in Walker doubling the reports of
aborting the walk. The irony is that in caller level (in this case, recreateAttributes method of AbstractRepository)
was properly handling this. It was only about improper logging -- spam.

Also, logging was improved, since by recent changes, this component got real Slf4j logger.