NFC is used by ALL repository types.
Even groups, even if in their case the actual _content_ of NFC is neglected, they still use it!
Before this change, the size of NFC in case of a non serviceable request to public group would raise
with 1+groupMemberSize. After this change, size of NFC would raise with the count of affected proxy
repositories only (1 in this case, Central only).