Implemented Walk throttling.
Walker component got "throttling" support in very non-intrusive way.
Everything works as before, but there is now possibility to implement
real "throttling" too, ie. intentionally slowing down the walks if needed.

This is currently used in AttributeUpgrade background thread only, as we detected
that thread priority has nothing to do with IO (but only with CPU) and that
presence of this thread might still saturate IO of a server (and leaving CPU
under utilized, hence, JVM does all "good" with this low prio thread).

But with this change, every "walk" is now capable of being throttled even without
(or with slight) code change.

The current implementation should be used as "last resort"! If we need this at all,
using the system property override (to set "throttle time in millis") could be used
to empirically "tune" and find best values for given (hence, similar) Nexus instance.

By default, upgrade thread is NOT throttled.