Fix for rare race condition.
I had this happened once, and might be related to what Alin reported this morning too:
the update indexes task started to run "out of the blue" of my just booted Nexus (virgin).
It was 1st start, so no configuration preexisted either.

What happened is due the async nature of this event inspector, and the fact
how it tried to discover "should the event be handled". Boot process does emit events
this inspector listens to! But, due to "burst" nature of boot, it is possible that the inspect
method executes when the process is actually done, but is about to process an event sent
during boot (is async). By moving that Nexus "state check" to the accept() method, that
is executed synchronously (!) we are shielded against this.