WalkerProcessor processItem method is invoked only of non-collection items
Historically, processItem was invoked for all kind of items, even collections.
As it turns out, we have NO WalkerProcessor that actually uses this, all of them
eagerly test for collections and simply do nothing. Another nice example is
SnapshotRemover, it's WalkerProcessor -- that IS interested in collections only
has empty processItem method (!), and performs all his work in onCollectionExit
method.