NEXUS-4789: Fixes
Culprit was HttpClientProxyUtil that was "touching" RemoteStorageContext on every
invocation. This was not an issue so far, since these calls happened only from
RemoteRepositoryStorages. But in a moment any "third party" component -- not
participating in RRS -- does the same, it causes all of RSS to detect as
"settings changed" and reconfigure their underlying transport.

Also, AHC provider modified to make use of newly introduced events
(thanks to Bentmann)

Added Javadocs to the two related events.