NEXUS-4845: Lowering loglevel to DEBUG in case of "cycles"
Because Nexus actually cannot have (cannot be brought in a state
by any means to have) cycles, and these messages are actually
only about doubled-processing of same repository while traversing
the request from a group (and groups of groups) recursively. Presence
of these repositories are still considered bad!

But, due to bug NXCM-3869 it we are actually the ones (staging promotion
more precisely) who are introducing these "double members", so putting
the level to DEBUG is fine for now, anyone using promotion would have
them anyway.