NEXUS-4787: Upgrade EHCache to latest
Currently it is 2.5.1. This change bumps EHCache version to latest,
but is not trivial one, especially since we have a legacy plexus
component used by both: nexus core and security. The sisu-ehcache
was cleaned, GAV changed to reflect these changes and is made a
drop-in replacement for the old plexus one. But, as new
pure SISU ehcache component depends on new AppContext, this
is bumped too (anyway needed for Jetty8 incoming changes).