Simplifying API.
The use of startup/shutdown makes user think these methods must be
called to have properly configured/running component. This is not the
case as everything happens in constructor. To lessen confusion,
startup is removed. This still makes usable the component if embedded
into UT as you just need to shutdown it, and create a new instance.