Another threading fix.
In rare cases, but Stuart's UT was failing on me with:

testCanStopAfterStartupFailure(org.sonatype.sisu.jetty.JettyConfigurationTest): We should not get here!
(line 252 in org.sonatype.sisu.jetty.JettyConfigurationTest)

The reason for this was simple: in Jetty8, if exception happened during server.start()
we immediately lowered the latch and only then filled in the exception value.
This meant, that in rare cases, the startJetty() method did get past of
started.await() latch, but the exception was still null, hence, the problem
went unnoticed. I had regular failures of this module (true, 1 out of 10
subsequent builds produced this, but it struck me this morning).

This change simple changes the assignment order: in case of starting exception,
we assign the exception first, and only then decrement the latch, ensuring
that letting loose the waiting thread on startJetty() will pick up the
exception, if any. This was an old problem, present even in old
Jetty7 class, not introduced by Stuart's changes.