NEXUS-4968: Simple fix.
It seems Java running on VM Windows platform is known to
"over sleep" but also "under sleep". Also, in links below
it is mentioned that "slice size on Windows" is 10ms, but might
be worse when running in Hypervisor.

This fix simply raises the sleep time over the "slice size" to
make it more plausible to "over sleep" instead to "under sleep".

See issue comments and links in them.