NEXUS-4968: Proper fix.
In short, not using "millisecond timestamp" for change detection.

In old code, the timestamp was not used at all to get the date
(except in one DEBUG log, but that was irrelevant, as change
date will be same as timestamp of logline it sits on!). The pure
reason of the timestamp was to support "change detection", but
it was wrong, as millisecond resolution is too coarse: think
concurrent changes, or just like in related issue, UT issues
with Windows running on Hypervisor and sleeping less than OS
slice and having JRE "under sleeping". See links in the
issue comments.