NXCM-3986: Improve logs
There is a complex set of conditions when to log what, when to log
stack traces and when not do log stack traces. This would need a fix
but for now, this change simply makes to log all HTTP 500 Internal
error responses _with_ stack trace, as previously they were logged
without it, and it did not make any sense.

This seems as long outstanding issue, as original commit implementing
this was unchanged for a long time, and it was causing that all
StorageException (that subclass IOException) gets logged _without_
stack trace, that is bad:

Change introduced in:
fcd4ccf26ad3e963d62e6e45ef06b8428f0d7608

We do have server side logs of the NXCM-3986, but as the cause was
not logged (no stack trace), we lost information why the rename
failed (clearly, it was some sort of FS related problem)