NXCM-3986: Removed duplicated if branches
Before, there was a separate handling of StorageException's
that was loosing stack trace. Also, superclass of StorageException
is IOException, and it was logged _with_ stack trace (in the else
branch). But both were creating a response of HTTP 500 internal error
(so, ultimately the only difference was the lack of the stack trace
in case of StorageException).

After this change, the StorageExceptions are handled in same way
as their superclass IOException, and will end up in logs _with_
stack trace too.