Fixes for attribute manager.
UID Attribute manager was "kicked in" too often, as it used
the NexusStateChangeEvent abstract event, and this event is
now parent class of 4 actual events, while the manager should be
"kicked" only on start or new plugin loaded/unloaded event.

Also, manager itself was using injected logger instead of the
helper AbstractLoggingComponent parent.