NEXUS-4948 improve test logging and remove CounterListener after done
We see on grid the testExpiration* methods are fragile, probably due to the fact it asserts against async event propagation.

Added logging that will likely prove this in the case that this test fails again.

Also each time the private doTestExpiration was called a new CounterListener was being added to application event multicaster and not removed for the next test call. This wasn't directly breaking tests, but it wasn't helping either - since the next test calling this method would add another CounterListener instance into the chain, slowing things down even more. So now we explicitly remove any CounterListener added when we are through.

Still don't like the fact the testExpiration* methods depend on eventing.