Security config changes.
While Nexus boots, Security was not loading configuration
as it was "lazy". Instead, the "upgrade" event inspector
was loading it eagerly, the one meant to perform upgrade.

So, a bit clean up, now security is pinged during nexus
boot (to force it config load), and upgrader inspector
made to kick in only when needed, not resulting in
flush/reload _always_, as it is not needed unless
upgrade happened.