NXCM-4024: Preparations.
Plugin POM and classpath was a mess, which actually came
from the fact that it inherited Nexus Parent, that made no sense
as it was full of Nexus specific depMgt and exclusion sections.

Now, the nexus maven plugins are "liberated" and use a clean
slate parent pom (the forge, same as Nexus parent uses) as parent.

Also, many cleanup and removals of unused dependencies happened.

And finally, as step to prepare NXCM-4024, the maven-core
already present dependency was "raised" from 2.2.1 to 3.0.4