Build fix of nexus-staging-plugin.
Aligning Aether version with one used in Nexus Core.
It made build break:
https://builds.sonatype.org/job/nexus-staging-plugin/38/

Because invoker "fills in" Core versioned aether (1.13) in local repo
while the verifier while running the plugin asks for other (1.13.1)
version of it.

Real solution is to bump core Aether version to 1.13.1 as it contains
relevant AETHER-127 fix.