NXCM-4381: Undoing goal name cleanup.
When work on Staging V2 started, the goals of the staging
related mojos were cleaned up, as some were doubled, and some
were inconsistently named.

To make the new release plugin 100% compat with existing builds
(as it should not change but emit the deprecation warning),
the cleanup is undone to not have goal name changes that would
break existing builds using the old goals.