Enhancing lock management in P2 plugin.
Minor line moves over all, but biggest change is in
AbstractP2MetadataSource where a lock.unlock() might "escape"
some finally block, as we had in NXCM-4391 (OOMError).

But, even if we have these changes, JVM does not guarantee, that
finally block will be invoked at all in case of Errors like OOM is,
but we try at least.