NEXUS-5104: P2 repository metadata update did not honor "localOnly"
P2 repository metadata update flag did not honor the localOnly flag
of requests, hence, it was possible to trigger md update even
over the UI by just browsing the repository.

(Equivalent would be for M2 repository, to have Nexus proxy and update
if artifact/metadata is old while you are browsing the Storage Tree
over UI, which never happens. The "/content" is different, and same
stands for P2, by "browsing" over "/content" updates/fetches WILL
be triggered for P2, just as is for M2 stuff)