Pull out FileContentLocator.
As P2 fixes proved, presence of "reusable implementation" of
File backed ContentLocator is good thing. The newly added class
to nexus-api module is 100% compatible with the one present in FS
(but that one was an "implementation details" of FS local storage)
and lacked some capabilities of handling tmp files.

P2 plugin for now holds it's own (copy) of this class in different
package, on next iteration it could be removed and change it to
use the newly added implementation.