NXCM-3986: Disabling indexing for test repository.
As reviewing the IT and code, there are no leaks or such. The
only explanation I could come up is indexer keeping the file
open (runs asynchronously in separate thread) and that was the
reason why file was not deletable on Win.

For now, Indexer is simply shut down, except in one IT that use it