NEXUS-5122: Simple yet workable protection against CME
We already had a "layer" of enhanced configuration that fixed
performance problems related to lookup and role resolutions.
Now, the same class has another role too, it ensures that code
retrieving configuration gets proper "snapshot in time", and hence
it's evaluation will be correct in the moment it got the config
but will not be affected by concurrent modifications of same
objects (role change introduced in a moment while role perms
are resolved for example).