"Detached" EnhancedConfiguration from Configuration.
It used to extend it, but not anymore to avoid any (possible) confusion.
Only one change needed. in config cleaner, and that change even makes it more
correct (it used the fact that was working on config object directly, using different
pattern that config manager -- not crud. So, this change it now aligns the two)