NXCM-4451: Fix NEXUS-5109: Reimplemented.
NEXUS-5109 initial implementation was wrong, as in some cases (repo
not added to registry, like RepositoryTypeDescriptor does) the
backing configuration did not exists. In those cases the plexus
lifecycle method #dispose() on AbstractRepository failed on it's
1st line (purge caches) with NPE, and hence the super.dispose()
was never invoked. That super.dispose is about to "untie" the
repository from event bus. Hence, the repository remained "dangling"
in there, without any configuration, and emitting NPE on every event
it was about to inspect.

NXCM-4451 is actually a proper reimplementation of NEXUS-5109 fix,
dispose method override from AbstractRepository is removed, as that is
"too late" point anyway (see above). A new event inspector is introduced
that listens for repo removals and cleans up the NFC. This means, that
in cases like the one with RepositoryTypeDescriptor this inspector is
not even invoked, and that is fine, since repositories created in it
are not even configured, and that was the main source of NPEs.