o Fix display of licensing detail page
Some layouts expect a BoxComponent (with setters for size and position).
The ExtJS documentation states that Ext.Button is a BoxComponent, but
the source actually extends only Ext.Component! Instead of "fixing" ext
js (and potentially breaking other buttons somewhere in the Nexus UI) we will
add the BoxComponent methods only to the browse button for now.