add profile for nexus-maven-plugins module
I had to remove the module to build the rest of nexus with aggregator pom and then nexus-maven-plugins separately, however we still need to build it on grid. This profile allows for that. Grid job will activate profile and build everything. And release plugin stays happy. And developer is caring less and less for this plugin...