NXCM-4516: Walker fix
When Walker invoked with starting path that was not pointing to
collection, it simply returned doing nothing. Otherwise, it would
recursively "dive" in.

This simply fix contains a small refactoring, pulling out
the "walkItem" part, and in main entry point, if the starting
item is not a collection, then direct call to this method is made.

UT added covering this case, that asserts that walk does happen
to the 1 targeted file item.