NEXUS-4692: Config version bumped and upgrade step added
Config version bumped from 2.0.0 to 2.0.1, and added upgrade step
that for now, only nullifies the localStorage provider field IF it is
set and is set to old apacheHttpClient3x.

In any other cases it does not do anything (as there might be
some manually tampered nexus.xml or what not).