NEXUS-5118: Fixing the initial imple
The initial implementation was conceived okay, but
there was a problem with Injection happening as the class
loaded up had annotations thrown away by classloader.

Reason for that is that in bundle layout the appender
is on Jetty classloader level (parent of Nexus webapp)
and that CL does not have @Inject. Hence, the fix was to
introduce "special" case for injection when ForwaringAppender
was met, simply doing manual injection on it.