NEXUS-5241: Simple fix to not have some events stored to Timeline.
Simply adding extra checks: event to be stored relates to a path
that is:

* NOT maven repository metadata file (maven-metadata.xml)
* NOT maven artifact signature file (*.asc)
* NOT maven checksum file (*.sha1 or *.md5)

Note: this change makes Timeline Maven-support dependant. Today, this is
fine as Maven support happens to be in Core, but once moved out, this
solution will either require to have Timeline plugin depend on Maven plugin
(which is wrong) or removal of this change will be needed.