NEXUS-5216: Fix
This issue fix was to NOT initiate a search when no context given (ctx is null).
When in Nexus, this corresponds to getting the archetype catalog of a repository
that has indexable=true set, hence, has no context at all. The bug was
that search was done, but due to empty list "non targeted search" was
made implicitly by Maven Indexer returning archetype from all other
repositories in Nexus but this one.