NEXUS-5249: UT ensuring proper behaviour.
For both cases: HTTP 404 (should be neglected), and any other IO
related exception (usually stemming from HTTP 401/403/50x or any other
problem, like storage related problems).

In both cases, it is validated that the processing of ALL repositories
is not stopped, and in second, that exception is thrown at the processing
end.