NEXUS-5249: Fixed the unstable nature of the UT
The changes that prepare() method did spawned some bg threads in core
to reindex apache-snapshots, and in same time the fetch counter
was installed. Due to concurrency, this sometimes catches (was installed
"in bad time") sometimes did not catch the call.

Now, after doing needed config changes, UT will wait for core to
"calm down" (tasks + events) and then continue to install the
counting components into DefaultIndexerManager.

This was the reason why "expected 1 but got 2" was reported, as same
apache-sbapshots was tried to be fetched twice actually: once by bg thread
and once by UT, if the method counter was installed in "bad time".