Simple change to make ITs work
Minimal change in AppContext story needed for ITs to work.

As Plexus tries to copy the passed in map (instead of delegating to it as SISU does),
now passing in the flattened AppContext as context.

Also, removed the if around nexus and nexus-test property file
loading, since they are optional (last ctor param is false, if not
found it will be silently ignored).