NEXUS-5280: ContextMemberProvider should always return same list of members.
At least during search processing is running. As it is now, it is directly
hooked onto group, that if group changed during search, causes to have
different repositories locked or unlocked. In case of member addition,
an IllegalMonitorStateEx would happen (a non-locked context would
be unlocked) and in case of member removal, the context of
removed member would remain locked.

There was an UT testing exactly this case, but for some reason it was commented
out. Test returned, code updated and added more checks and explanations
in comment.

Method setRepositoryIndexContextSearchable() was called when it
was actually not needed.