NEXUS-5280: Fixing "look ahead" problem.
With NEXUS-5280 implemented in a way it was done originally, Nexus
lost the ability to load up configuration where a group G1
referenced G2 as member, while G2 was not YET loaded (is a valid
scenario).

This is fixed by not

Note: as it is obvious, doing GroupRepository#getMemberRepositories()
during boot (since this code is invoked during boot, but also during
"normal" runtime reconfigurations) results in NoSuchRepoEx that made
many "group of groups" ITs to fail. Reason is that method above goes
to RepositoryRegistry to "resolve" member IDs to Repository instances,
but G2 is not yet in registry.

Solution is to make lookup "lazy": similarly as before, now the
list of member IDs are "remembered", and the list of contexts is
lazily constructed. In cases like member changes or disabling
indexing on a repository, IndexerManager will recreate
LazyContextMemberProvider just it did before.