allow running (JUnit) tests named Abstract* that are not 'abstract'
We have in codebase test classes named 'Abstract*' but are not really 'abstract' and ones that are really abstract ( meant to be extended only )

surefire will do the right thing by allowing it to try to run any class that ends in Test and starts with Abstract. This will allow us to have test for real classes that are named Abstract* and keep the naming convention that IDEs expect for tests.

'abstract' test classes are still not run. If developer intends otherwise to execute a test class named 'Abstract*' then add one or more test methods and do not make it abstract.

I noticed several tests that were  being missed besides the ones covered in nexus-5197 because once upon a time we excluded Abstract named classes. ( probably my suggestion ).

I've left failsafe config to ignore Abstract* named test classes because I am unclear on TestNG semantics and also those should be integration tests, which do not normally follow the convention of unit tests where tests are named one to one to the unit under test + 'Test' suffix